+++ src/motion/CoreXYKinematics.hpp
@@ -4,23 +4,23 @@
 #include <cstdint>
 
 struct MotorSteps {
-    float a;
-    float b;
+    double a;
+    double b;
 };
 
 struct XYPos {
-    float x_mm;
-    float y_mm;
+    double x_mm;
+    double y_mm;
 };
 
 class CoreXYKinematics {
 public:
-    explicit CoreXYKinematics(float steps_per_mm_);
+    explicit CoreXYKinematics(double steps_per_mm_);
 
     MotorSteps mm_to_steps(const XYPos& pos) const;
     XYPos steps_to_mm(const MotorSteps& steps) const;
 
 private:
-    float steps_per_mm;
+    double steps_per_mm;
 };
 #endif
